Add some graph traversal algorithms
Depth first search (DFS), breadth first search (BFS) and Prim's algorithm
have a similar structure. All three create a tree from a graph that
visits each node. Prim's algorithm is a greedy algorithm that creates a
minimum spanning tree of the graph.

Currently there isn't any animation. The graph is generated randomly
which is sometimes not connected. Maybe there is a better algorithm for
that :)